Add files via upload
Excel Case Study 

Analytics allow you to quantify the effects of making a change to your strategy, and that's invaluable to the process of improving and optimizing campaigns. The biggest benefit of utilizing proper analytics is being able to identify strengths and weaknesses. Problem statements are mentioned along with the datasets in the given excel sheets. Please download the file and answer the questions.

● Create a 3-dimensional column chart comparing sales data for men and women, but omitting BMWs

● Create a chart to compare the favorite films data for 15-25 years old only (be careful not to include any unnecessary blank rows or columns in your selected data).

Format this chart so that it is a pie chart, with the Barbarella slice "exploded" and each segment labeled:

● Select the necessary ranges of data to create a 3-D cone chart showing the City and the Population

● Convert this data into a line chart. Make the necessary changes to ensure that it resembles the one shown below.

● Create a chart which shows the top 6 countries and their medal hauls

● Plot a graph to compare the sales of 2012 and 2013. Also, show the growth on the same page.